Modified the buildable level values to add colorRent.
Make the checks for player and bank to see if each can pay.

Made the functions for each of the chance and community chest cards and shortened the dictionaries for each.

Started the consultancy fee, beauty contest, and inherit functions. They need checks to see if the bank can provide the funds still.

Finished the street repair function and documented it.

commented some of the original variables and reorginized the location of them.