Modified some of the parameter and variable names to get rid of some of the warnings.
modified the playerbankrequest function so that it works for any withdraw from players wallet.

Added a function for the dice rolls

Changed advance functions location variable to new_location for clearer understanding throughout the program
Documented advance function as well.

Made the railroad function and documented it.
Made the utility function and documented it.
Need to change these two function names so that they refer to the chance cards and nothing else.
Still trying to figure out how to condense Railroad function and Utility function as there is a LOT of similar code.

fixed the community_chest_cards name to draw_cards.